refactor: SocialAccount & EmailAddress components
