+++ src/components/[guild]/JoinModal/components/ConnectPlatform.tsx
@@ -3,7 +3,7 @@ import { useWeb3React } from "@web3-react/core"
 import useAccess from "components/[guild]/hooks/useAccess"
 import usePlatformsToReconnect from "components/[guild]/hooks/usePlatformsToReconnect"
 import useUser from "components/[guild]/hooks/useUser"
-import { TwitterV1Tooltip } from "components/common/Layout/components/Account/components/AccountModal/components/SocialAccount"
+import { TwitterV1Tooltip } from "components/common/Layout/components/Account/components/AccountModal/components/SocialAccount/SocialAccount"
 import Script from "next/script"
 import platforms from "platforms/platforms"
 import { useEffect } from "react"
