+++ src/components/common/Layout/components/Account/components/AccountModal/components/SocialAccount/EmailAddress.tsx
@@ -1,0 +1,272 @@
+import {
+  Collapse,
+  FormControl,
+  FormLabel,
+  HStack,
+  IconButton,
+  Input,
+  ModalBody,
+  ModalCloseButton,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  ModalOverlay,
+  PinInput,
+  PinInputField,
+  Text,
+  Tooltip,
+  VStack,
+  useDisclosure,
+} from "@chakra-ui/react"
+import { useConnectEmail } from "components/[guild]/JoinModal/hooks/useConnectPlatform"
+import processJoinPlatformError from "components/[guild]/JoinModal/utils/processJoinPlatformError"
+import useUser from "components/[guild]/hooks/useUser"
+import Button from "components/common/Button"
+import { Error } from "components/common/Error"
+import { Modal } from "components/common/Modal"
+import { SignedValdation, useSubmitWithSign } from "hooks/useSubmit"
+import useToast from "hooks/useToast"
+import { PencilSimple } from "phosphor-react"
+import platforms from "platforms/platforms"
+import { useEffect, useRef, useState } from "react"
+import { useController, useForm, useWatch } from "react-hook-form"
+import fetcher from "utils/fetcher"
+import { useDisconnectEmail } from "../../hooks/useDisconnect"
+import DisconnectAccountButton from "./components/DisconnectAccountButton"
+import SocialAccountUI from "./components/SocialAccountUI"
+
+const EmailAddress = () => {
+  const { emails } = useUser()
+  const isConnected = !!emails && !emails.pending
+
+  return (
+    <SocialAccountUI
+      type={"EMAIL"}
+      username={emails?.emailAddress}
+      isConnected={isConnected}
+    >
+      {isConnected ? <ConnectEmailButton /> : <DisconnectEmailButton />}
+    </SocialAccountUI>
+  )
+}
+
+const PIN_LENGTH = 6
+const TIMEOUT = 10_000
+
+const ConnectEmailButton = () => {
+  const { emails } = useUser()
+  const { isOpen, onOpen, onClose } = useDisclosure()
+  const [pendingEmailAddress, setPendingEmailAddress] = useState(
+    emails?.pending ? emails?.emailAddress : null
+  )
+
+  const { register, handleSubmit, control, setValue, reset } = useForm<{
+    email: string
+    code: string
+  }>({
+    defaultValues: { email: "", code: "" },
+  })
+  const { field } = useController({ control, name: "code" })
+  const email = useWatch({ control, name: "email" })
+  const { id: userId } = useUser()
+  const toast = useToast()
+
+  const [emailSentAt, setEmailSentAt] = useState<number>(null)
+
+  const submitVerificationRequest = (
+    signedPayload: SignedValdation
+  ): Promise<{ remainingAttempts: number; success: boolean }> =>
+    fetcher(`/v2/users/${userId}/emails`, signedPayload).then((data) => {
+      setEmailSentAt(Date.now())
+      return data
+    })
+
+  const verificationRequest = useSubmitWithSign(submitVerificationRequest)
+
+  const shouldShowPinEntry = !!verificationRequest.response || !!pendingEmailAddress
+
+  const handleOnClose = () => {
+    const channel = new BroadcastChannel("EMAIL")
+    // Sending null, so useOauthPopup's promise resolves, but we won't send a connect in useConnectPlatform because authData is falsy
+    channel.postMessage({ type: "OAUTH_SUCCESS", data: null })
+    channel.close()
+    verificationRequest.reset()
+    reset()
+    if (emails?.emailAddress) {
+      setPendingEmailAddress(emails.emailAddress)
+      setValue("email", emails.emailAddress)
+    }
+    onClose()
+  }
+
+  const submit = handleSubmit((formValues) =>
+    verificationRequest.onSubmit({ emailAddress: formValues.email })
+  )
+
+  const pinInputRef = useRef<HTMLInputElement>()
+
+  const connect = useConnectEmail({
+    onSuccess: () => {
+      toast({ status: "success", title: "Email verified" })
+      handleOnClose()
+    },
+    onError: () => {
+      setValue("code", "")
+      pinInputRef.current?.focus()
+    },
+  })
+
+  // Set the known pending email address
+  useEffect(() => {
+    if (!emails?.pending) return
+    setPendingEmailAddress(emails?.emailAddress)
+    setValue("email", emails?.emailAddress)
+  }, [emails])
+
+  // Timer to decide if resend button is disabled
+  const [dateNow, setDateNow] = useState(Date.now())
+  const canResend = dateNow - emailSentAt > TIMEOUT
+  useEffect(() => {
+    const interval = setInterval(() => setDateNow(Date.now()), TIMEOUT)
+    return () => clearInterval(interval)
+  }, [])
+  const isResendButtonDisabled = shouldShowPinEntry && !canResend && !connect.error
+
+  return (
+    <>
+      <Button
+        onClick={onOpen}
+        colorScheme={emails?.pending ? "orange" : platforms.EMAIL.colorScheme}
+        variant={"solid"}
+        size="sm"
+      >
+        {emails?.pending ? "Verify" : "Connect"}
+      </Button>
+      <Modal isOpen={isOpen} onClose={handleOnClose} size="sm">
+        <ModalOverlay />
+        <ModalContent>
+          <form onSubmit={submit}>
+            <ModalHeader>Connect email</ModalHeader>
+            <ModalCloseButton />
+            <ModalBody display={"flex"} flexDir="column" pb="6">
+              <Error
+                error={verificationRequest.error ?? connect.error}
+                processError={processJoinPlatformError}
+              />
+
+              <Collapse
+                in={!shouldShowPinEntry}
+                unmountOnExit
+                style={{ padding: "1px" }}
+              >
+                <FormControl>
+                  <FormLabel>Email address</FormLabel>
+                  <Input
+                    autoFocus
+                    type="email"
+                    placeholder="me@example.com"
+                    {...register("email", {
+                      required: true,
+                      // We should get native validation
+                      // pattern: { value: EMAIL_REGEX, message: "Invalid E-Mail address" },
+                    })}
+                  />
+                </FormControl>
+              </Collapse>
+              <Collapse
+                in={shouldShowPinEntry}
+                unmountOnExit
+                style={{ padding: "1px" }}
+              >
+                <VStack spacing={4}>
+                  <Text textAlign={"center"}>
+                    Enter the code we've sent to {email}{" "}
+                    <IconButton
+                      size={"xs"}
+                      icon={<PencilSimple />}
+                      aria-label="Use different email address"
+                      onClick={() => {
+                        setPendingEmailAddress(null)
+                        setValue("code", "")
+                        setValue("email", "")
+                        connect.reset()
+                        verificationRequest.reset()
+                      }}
+                    />
+                  </Text>
+                  <HStack justifyContent={"center"}>
+                    <PinInput
+                      isInvalid={connect.error}
+                      value={field.value}
+                      onChange={(value) => {
+                        field.onChange(value)
+                        if (value.length === PIN_LENGTH) {
+                          connect.onSubmit({
+                            authData: { code: value },
+                            emailAddress: email,
+                          })
+                        }
+                      }}
+                    >
+                      <PinInputField autoFocus ref={pinInputRef} />
+                      <PinInputField />
+                      <PinInputField />
+                      <PinInputField />
+                      <PinInputField />
+                      <PinInputField />
+                    </PinInput>
+                  </HStack>
+                </VStack>
+              </Collapse>
+            </ModalBody>
+            <ModalFooter>
+              <Tooltip
+                label={
+                  "Check if you received the first code before requesting a new one (check in the spam as well)"
+                }
+                hasArrow
+                isDisabled={!isResendButtonDisabled}
+              >
+                <Button
+                  w="full"
+                  isLoading={
+                    verificationRequest.isLoading ||
+                    connect.isLoading ||
+                    connect.isSigning
+                  }
+                  size={shouldShowPinEntry ? "sm" : undefined}
+                  colorScheme={shouldShowPinEntry ? "gray" : "green"}
+                  variant={shouldShowPinEntry ? "ghost" : "solid"}
+                  type={"submit"}
+                  isDisabled={isResendButtonDisabled}
+                >
+                  {shouldShowPinEntry ? "Resend code" : "Send code"}
+                </Button>
+              </Tooltip>
+            </ModalFooter>
+          </form>
+        </ModalContent>
+      </Modal>
+    </>
+  )
+}
+
+const DisconnectEmailButton = () => {
+  const disclosure = useDisclosure()
+  const { emails } = useUser()
+
+  const { onSubmit, isLoading, signLoadingText } = useDisconnectEmail(
+    disclosure.onClose
+  )
+  const onConfirm = () => onSubmit(emails?.emailAddress)
+  const loadingText = signLoadingText ?? "Removing"
+
+  return (
+    <DisconnectAccountButton
+      name={platforms.EMAIL.name}
+      {...{ disclosure, isLoading, loadingText, onConfirm }}
+    />
+  )
+}
+
+export default EmailAddress
