+++ src/components/common/Layout/components/Account/components/AccountModal/components/SocialAccount/index.ts
@@ -1,0 +1,5 @@
+import EmailAddress from "./EmailAddress"
+import SocialAccount from "./SocialAccount"
+
+export default SocialAccount
+export { EmailAddress }
